[KOGITO-1373] Create new package to handle state control
