+++ packages/editor-state-control-manager/README
@@ -1,0 +1,1 @@
+The propose of this package is to save the events that are done on the editor so it give helper methods to change (Undo/Redo operation) or show (isDirty indicator) the current state.
