+++ packages/online-editor/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@kogito-tooling/core-api": "0.4.3",
+    "@kogito-tooling/editor-state-control-manager": "0.4.3",
     "@kogito-tooling/embedded-editor": "0.4.3",
     "@kogito-tooling/patternfly-base": "0.4.3",
     "@octokit/rest": "16.35.0",
