+++ packages/online-editor/src/__tests__/testing_utils.tsx
@@ -22,7 +22,7 @@ import { HashRouter } from "react-router-dom";
 import { GithubService } from "../common/GithubService";
 import { GlobalContext, GlobalContextType } from "../common/GlobalContext";
 import { Routes } from "../common/Routes";
-import { StateControl } from "../editor/StateControl";
+import { StateControl } from "@kogito-tooling/editor-state-control-manager";
 
 export function usingTestingGlobalContext(children: React.ReactElement, ctx?: Partial<GlobalContextType>) {
   const usedCtx = {
